refactor: domain 설정 수정

+++ src/main/java/com/wanted/budgetmanagement/domain/budget/entity/Budget.java
@@ -34,7 +34,7 @@ public class Budget {
     private BudgetCategory category;
 
     @Column
-    private int money;
+    private long money;
 
     @Temporal(TemporalType.DATE)
     private Date period;
